+++ src/authentication/defaultSessionManager.ts
@@ -1,20 +1,21 @@
-export interface SessionManager {
-    /**
-     * Returns stored value.
-     * @param key 
-     */
-    getItem<T>(key: string): Promise<T>;
+import { SessionManager } from "./sessionManager";
 
-    /**
-     * Stores value with specified key.
-     * @param key {string} Stored value key.
-     * @param value {T} Stored value.
-     */
-    setItem<T>(key: string, value: T): Promise<void>;
+export class DefaultSessionManager implements SessionManager {
+    public async getItem<T>(key: string): Promise<T> {
+        const value = sessionStorage.getItem(key);
 
-    /**
-     * Removed value with specified key.
-     * @param key {string} Stored value key.
-     */
-    removeItem(key: string): Promise<void>;
+        if (!value) {
+            return null;
+        }
+
+        return JSON.parse(value);
+    }
+
+    public async setItem<T>(key: string, value: T): Promise<void> {
+        sessionStorage.setItem(key, JSON.stringify(value));
+    }
+
+    public async removeItem(key: string): Promise<void> {
+        sessionStorage.removeItem(key);
+    }
 }
